+++ lib/domain/models/settings.dart
