+++ lib/ui/pages/home/home.dart
@@ -1,10 +1,10 @@
 import 'package:flutter/material.dart';
 import 'package:get_it/get_it.dart';
-import 'package:yss_todo/controllers/home.dart';
+import 'package:yss_todo/domain/controllers/home.dart';
 import 'package:yss_todo/helpers.dart';
 import 'package:yss_todo/logger.dart';
-import 'package:yss_todo/pages/home/widgets/appbar.dart';
-import 'package:yss_todo/pages/home/widgets/tasklist.dart';
+import 'package:yss_todo/ui/pages/home/widgets/appbar.dart';
+import 'package:yss_todo/ui/pages/home/widgets/tasklist.dart';
 
 class Homepage extends StatelessWidget {
   const Homepage({super.key});
